cnchi/src/Cnchi-master/requirements.txt
pyalpm>=0.7
pygobject>=3.16.0
pydbus
pycairo
pyparted>=3.10.1
mako
requests
